Rename PyFuncEnsembler to PyFuncEnsemblerRunner to remove overloaded classname
