--- engines/live-ensembler/pyfunc_ensembler/__init__.py
@@ -1,2 +1,0 @@
-from .ensembler import PyFuncEnsembler
-
