+++ engines/live-ensembler/pyfunc_ensembler_runner/__init__.py
@@ -1,0 +1,2 @@
+from .ensembler_runner import PyFuncEnsembler
+
