+++ engines/live-ensembler/pyfunc_ensembler_runner/ensembler_runner.py
@@ -1,14 +1,13 @@
 import pandas as pd
+
+from turing.ensembler import PyFunc
 from mlflow import pyfunc
 
 
-class PyFuncEnsembler:
+class PyFuncEnsemblerRunner:
     """
-    PyFunc ensembler used for real-time outputs
+    PyFunc ensembler runner used for real-time outputs
     """
-    FEATURE_COLUMN_PREFIX = '__features__'
-    PREDICTION_COLUMN_PREFIX = '__predictions__'
-    TREATMENT_CONFIG_COLUMN_PREFIX = '__treatment_config__'
 
     def __init__(self, artifact_dir: str):
         self.artifact_dir = artifact_dir
@@ -20,45 +19,46 @@ class PyFuncEnsembler:
         self.ready = True
 
     def predict(self, inputs: dict) -> dict:
-        ensembler_inputs = PyFuncEnsembler.preprocess_input(inputs)
+        ensembler_inputs = PyFuncEnsemblerRunner.preprocess_input(inputs)
         output = self._ensembler.predict(ensembler_inputs).iloc[0].to_list()
         return output
 
     @staticmethod
     def preprocess_input(inputs: dict):
-        """
-        Dummy preprocessing method
-        """
-        features = pd.Series(PyFuncEnsembler._get_features_from_inputs(inputs))
-        predictions = pd.Series(PyFuncEnsembler._get_predictions_from_inputs(inputs))
-        treatment_config = pd.Series(PyFuncEnsembler._get_treatment_config_from_inputs(inputs))
+        features = pd.Series(PyFuncEnsemblerRunner._get_features_from_inputs(inputs))
+        predictions = pd.Series(PyFuncEnsemblerRunner._get_predictions_from_inputs(inputs))
+        treatment_config = pd.Series(PyFuncEnsemblerRunner._get_treatment_config_from_inputs(inputs))
         preprocessed_input = pd.concat([features, predictions, treatment_config]).to_frame().transpose()
         return preprocessed_input
 
     @staticmethod
     def _get_features_from_inputs(inputs: dict) -> dict:
-        raw_predictions = PyFuncEnsembler._flatten_json(inputs['request'])
-        features = PyFuncEnsembler._create_dict_with_headers(raw_predictions,
-                                                             PyFuncEnsembler.FEATURE_COLUMN_PREFIX)
+        features = PyFuncEnsemblerRunner._flatten_json(inputs['request'])
         return features
 
     @staticmethod
     def _get_predictions_from_inputs(inputs: dict) -> dict:
-        raw_predictions = PyFuncEnsembler._flatten_json(inputs['response']['route_responses'])
-        predictions = PyFuncEnsembler._create_dict_with_headers(raw_predictions,
-                                                                PyFuncEnsembler.PREDICTION_COLUMN_PREFIX)
+        raw_predictions = PyFuncEnsemblerRunner._flatten_json(inputs['response']['route_responses'])
+        predictions = PyFuncEnsemblerRunner._create_dict_with_headers(raw_predictions,
+                                                                      PyFunc.PREDICTION_COLUMN_PREFIX)
         return predictions
 
     @staticmethod
     def _get_treatment_config_from_inputs(inputs: dict) -> dict:
-        raw_predictions = PyFuncEnsembler._flatten_json(inputs['response']['experiment'])
+        raw_predictions = PyFuncEnsemblerRunner._flatten_json(inputs['response']['experiment'])
 
-        treatment_config = PyFuncEnsembler._create_dict_with_headers(raw_predictions,
-                                                                     PyFuncEnsembler.TREATMENT_CONFIG_COLUMN_PREFIX)
+        treatment_config = PyFuncEnsemblerRunner._create_dict_with_headers(raw_predictions,
+                                                                           PyFunc.TREATMENT_CONFIG_COLUMN_PREFIX)
         return treatment_config
 
     @staticmethod
     def _flatten_json(y):
+        """
+        Helper function to normalise a nested dictionary into a dictionary of depth 1 with names following the
+        convention: key_1.key_2.key_3..., with a period acting as a delimiter between nested keys
+
+        Items in lists have their names rendered using their index numbers within the lists they are found in.
+        """
         out = {}
 
         def flatten(x, name=''):
